refactor(install): extract the UI-agnostic install engine into internal/installer
Phase 4 step 2 (tview-to-Charm migration), behavior-neutral by
construction: the characterization goldens pass unchanged.

Moved verbatim from internal/tui/wizard: the wizard data model
(InstallWizardPrefill, InstallWizardData), the single config template
builder (ApplyInstallData with its secondary/cloud validators), prefill
derivation (DeriveInstallWizardPrefill, ParseEnvTemplate), template key
helpers, email delivery method normalization, and the post-install audit
collector (audit.go, with its test suite).

The surviving tview wizard screens keep compiling through a thin alias
bridge (installer_aliases.go) that dies with them in the next step; the
CLI now consumes internal/installer directly for prefill and audit
collection.
