+++ cmd/proxsave/install.go
@@ -15,6 +15,7 @@ import (
 	cronutil "github.com/tis24dev/proxsave/internal/cron"
 	"github.com/tis24dev/proxsave/internal/identity"
 	"github.com/tis24dev/proxsave/internal/logging"
+	"github.com/tis24dev/proxsave/internal/installer"
 	"github.com/tis24dev/proxsave/internal/safeexec"
 	"github.com/tis24dev/proxsave/internal/tui/wizard"
 	buildinfo "github.com/tis24dev/proxsave/internal/version"
@@ -175,7 +176,7 @@ func runPostInstallAuditCLI(ctx context.Context, reader *bufio.Reader, execPath,
 		bootstrap.Info("Post-install audit: running dry-run (this may take a minute)")
 	}
 
-	suggestions, err := wizard.CollectPostInstallDisableSuggestions(ctx, execPath, configPath)
+	suggestions, err := installer.CollectPostInstallDisableSuggestions(ctx, execPath, configPath)
 	if err != nil {
 		fmt.Printf("WARNING: Post-install check failed (non-blocking): %v\n", err)
 		if bootstrap != nil {
@@ -626,7 +627,7 @@ func configureSecondaryStorage(ctx context.Context, reader *bufio.Reader, templa
 	fmt.Println("Network shares must be mounted BEFORE running this backup tool.")
 	fmt.Println("For direct network access without mounting, use cloud storage (rclone) instead.")
 	fmt.Println("(You can change these settings later in backup.env)")
-	prefill := wizard.DeriveInstallWizardPrefill(template)
+	prefill := installer.DeriveInstallWizardPrefill(template)
 	enableSecondary, err := confirmDefault(ctx, reader, "Enable secondary backup path?", prefill.SecondaryEnabled)
 	if err != nil {
 		return "", err
@@ -668,7 +669,7 @@ func configureSecondaryStorage(ctx context.Context, reader *bufio.Reader, templa
 func configureCloudStorage(ctx context.Context, reader *bufio.Reader, template string) (string, error) {
 	fmt.Println("\n--- Cloud storage (rclone) ---")
 	fmt.Println("Remember to configure rclone manually before enabling cloud backups.")
-	prefill := wizard.DeriveInstallWizardPrefill(template)
+	prefill := installer.DeriveInstallWizardPrefill(template)
 	enableCloud, err := confirmDefault(ctx, reader, "Enable cloud backups?", prefill.CloudEnabled)
 	if err != nil {
 		return "", err
@@ -699,7 +700,7 @@ func configureFirewallRules(ctx context.Context, reader *bufio.Reader, template 
 	fmt.Println("\n--- Firewall rules ---")
 	fmt.Println("Enable collection of firewall rules (e.g., iptables/nftables).")
 	fmt.Println("(You can change this later in backup.env via BACKUP_FIREWALL_RULES)")
-	enable, err := confirmDefault(ctx, reader, "Backup firewall rules?", wizard.DeriveInstallWizardPrefill(template).FirewallEnabled)
+	enable, err := confirmDefault(ctx, reader, "Backup firewall rules?", installer.DeriveInstallWizardPrefill(template).FirewallEnabled)
 	if err != nil {
 		return "", err
 	}
@@ -712,7 +713,7 @@ func configureFirewallRules(ctx context.Context, reader *bufio.Reader, template 
 }
 
 func configureNotifications(ctx context.Context, reader *bufio.Reader, template string) (string, error) {
-	prefill := wizard.DeriveInstallWizardPrefill(template)
+	prefill := installer.DeriveInstallWizardPrefill(template)
 	fmt.Println("\n--- Telegram ---")
 	enableTelegram, err := confirmDefault(ctx, reader, "Enable Telegram notifications (centralized)?", prefill.TelegramEnabled)
 	if err != nil {
@@ -781,7 +782,7 @@ func promptEmailDeliveryMethod(ctx context.Context, reader *bufio.Reader, defaul
 
 func configureEncryption(ctx context.Context, reader *bufio.Reader, template *string) (bool, error) {
 	fmt.Println("\n--- Encryption ---")
-	enableEncryption, err := confirmDefault(ctx, reader, "Enable backup encryption?", wizard.DeriveInstallWizardPrefill(*template).EncryptionEnabled)
+	enableEncryption, err := confirmDefault(ctx, reader, "Enable backup encryption?", installer.DeriveInstallWizardPrefill(*template).EncryptionEnabled)
 	if err != nil {
 		return false, err
 	}
